refactor: extract data to files
Also general improvements on BioCard.
