+++ home/src/App.tsx
@@ -4,20 +4,12 @@ import Header from "./components/Header";
 import Home from "./components/Home";
 //import Footer from "./components/Footer";
 
-
-
 export default function App() {
-
-
   return (
     <div>
-      <header>
-        <Header />
-      </header>
+      <Header />
       <body>
         <Home />
-
-
       </body>
     </div>
   );
