+++ home/src/components/Articles.tsx
@@ -1,11 +1,4 @@
 
-import Accordion from "@mui/material/Accordion";
-import AccordionSummary from "@mui/material/AccordionSummary";
-import AccordionDetails from "@mui/material/AccordionDetails";
-import Typography from "@mui/material/Typography";
-
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-
 import Timeline from "./Timeline";
 import Pill from "./Pill";
 import SectionHeader from "./SectionHeader";
