+++ home/src/components/Home.tsx
@@ -9,14 +9,8 @@ import { applyTheme } from "../themes/utils";
 import baseTheme from "../themes/base";
 import darkTheme from "../themes/dark";
 
-//import portrait from "../images/mrodriguez-portrait-small.jpg";
-//import portraitDark from "../images/csharp.png";
-
 import BioCard from "./BioCard";
 import SkewedCard from "./SkewedCard";
-
-import Skeletor from "./skeletons/Skeletor";
-
 
 export default function Home() {
   useEffect(() => {
@@ -25,22 +19,14 @@ export default function Home() {
 
   return (
     <div>
+      <BioCard />
 
-      <div>
-        <BioCard />
-      </div>
-
-      <Skeletor />
       <div className="bg-primary-light py-10">
         <Articles />
       </div>
 
-      <SkewedCard text="Woah. What do you think about this?" />
-
-
-
       <StickyFooter />
-      {/* <Footer /> */}
+
     </div>
   );
 }
