+++ home/src/components/SubSectionHeader.tsx
@@ -1,0 +1,10 @@
+import "../index.css";
+
+export default function SubSectionHeader({ title }) {
+  return (
+    <div
+      className="bg-secondary text-text-on-secondary font-bold rounded px-4 py-1 mb-1">
+      {title}
+    </div>
+  );
+}
