+++ home/src/components/WorkExperienceCard.tsx
@@ -1,6 +1,7 @@
 import "../index.css";
 import Pill from "./Pill";
 import WorkExperienceDetailCard from "./WorkExperienceDetailCard";
+import SubSectionHeader from "./SubSectionHeader";
 
 // continuar aqui, y que reciba las pills y details desde Timeline.
 export default function WorkExperienceCard({ company, title, dateStart, dateEnd, pills, details }) {
@@ -8,9 +9,9 @@ export default function WorkExperienceCard({ company, title, dateStart, dateEnd,
 
   return (
     <div className="order-1 rounded-lg w-full px-4 py-4 border mb-4 border-primary-dark bg-primary">
-      <p className="mb-2 text-xl font-bold text-secondary">{company}</p>
+      <SubSectionHeader title={company} />
 
-      {/* pills start */}
+      {/* pills */}
       <div className="container mx-auto my-1 p-0 flex flex-wrap">
         {
           pills.map((item, index) => (
@@ -18,7 +19,6 @@ export default function WorkExperienceCard({ company, title, dateStart, dateEnd,
           ))
         }
       </div>
-      {/* pills end */}
 
       <p className="text-text-on-secondary">
         <WorkExperienceDetailCard
