+++ home/src/components/data/pills_arconsultores.tsx
@@ -1,0 +1,7 @@
+const pills_arconsultores = [
+    { name: "asp classic", },
+    { name: "SQL Server", },
+    { name: "Web Services", },
+  ];
+
+  export default pills_arconsultores;
