+++ home/src/components/data/pills_bank_cba.tsx
@@ -1,0 +1,8 @@
+const pills_bank_cba = [
+    { name: "C#", },
+    { name: "SQL Server", },
+    { name: "Oracle", },
+    { name: "vb.net", },
+  ];
+
+export default pills_bank_cba;
