+++ home/src/components/data/pills_exet.tsx
@@ -1,0 +1,7 @@
+const pills_exet = [
+    { name: "Oracle", },
+    { name: "Service Bus", },
+    { name: "SQL Server", },
+  ];
+
+export default pills_exet;
