+++ home/src/components/data/pills_flux.tsx
@@ -1,0 +1,7 @@
+const pills_flux = [
+    { name: '.netcore', },
+    { name: 'KeyVault', },
+    { name: 'Microservices', },
+  ];
+
+export default pills_flux;
