+++ home/src/components/data/pills_gruppoit.tsx
@@ -1,0 +1,8 @@
+const pills_gruppoit = [
+    { name: "asp.net", },
+    { name: "SQL Server", },
+    { name: "GDI+", },
+    { name: "XML", },
+  ];
+
+export default pills_gruppoit;
