+++ home/src/components/data/pills_hka.tsx
@@ -1,0 +1,7 @@
+const pills_hka = [
+    { name: "asp classic", },
+    { name: "SQL Server", },
+    { name: "Web Services", },
+  ];
+
+  export default pills_hka;
