+++ home/src/components/data/pills_hp.tsx
@@ -1,0 +1,8 @@
+  const pills_hp = [
+    { name: "C#", },
+    { name: "SQL Server", },
+    { name: "J2EE", },
+    { name: "ETL", },
+  ];
+
+  export default pills_hp;
