+++ home/src/components/data/pills_registro.tsx
@@ -1,0 +1,7 @@
+const pills_registro = [
+    { name: "C#", },
+    { name: "WPF", },
+    { name: "C wrappers", },
+  ];
+
+export default pills_registro;
