+++ home/src/components/data/pills_southworks.tsx
@@ -1,0 +1,8 @@
+const pills_southworks = [
+    { name: '.netcore', },
+    { name: 'azure', },
+    { name: 'React', },
+    { name: 'React', },
+  ];
+
+  export default pills_southworks;
