+++ home/src/components/data/pills_tg.tsx
@@ -1,0 +1,8 @@
+const pills_tg = [
+    { name: '.netcore', },
+    { name: 'react', },
+    { name: 'azure', },
+    { name: 'tailwind css', },
+  ];
+
+export default pills_tg;
