+++ home/src/components/data/pills_utn.tsx
@@ -1,0 +1,8 @@
+export const pills_utn = [
+  { name: "J2EE" },
+  { name: "IBM DB/2" },
+  { name: "WebSphere" },
+  { name: "XML" },
+];
+
+export default pills_utn;
