+++ home/src/components/data/workExperience_arconsultores.tsx
@@ -1,0 +1,10 @@
+const workExperience_arconsultores = [
+    { desc: "◾ Client: Claro (El Salvador, Central America)" },
+    { desc: "Consulting / developing for telecom provider." },
+    { desc: "Traveled to work in El Salvador (1 month)." },
+    {
+      desc: "Provided onsite consulting about .NET / Oracle, for the ERP system of the company.",
+    },
+  ];
+
+export default workExperience_arconsultores;
