+++ home/src/components/data/workExperience_bank_cba.tsx
@@ -1,0 +1,10 @@
+const workExperience_bank_cba = [
+    {
+      desc: "Financial software maintenance: Credit cards, Loans, Investments, fraud.",
+    },
+    {
+      desc: "Performed several tasks of maintenance of existing applications, and backend operations.",
+    },
+  ];
+
+export default workExperience_bank_cba;
