+++ home/src/components/data/workExperience_exet.tsx
@@ -1,0 +1,11 @@
+const workExperience_exet = [
+    { desc: "Project RentUp" },
+    {
+      desc: "Client: General Income Directorate (DGR), Córdoba, Ministry of Modernization.",
+    },
+    {
+      desc: "Tasks: Analysis, investigation, interviews, design, programming.",
+    },
+  ];
+
+  export default workExperience_exet;
