+++ home/src/components/data/workExperience_flux.tsx
@@ -1,0 +1,10 @@
+const workExperience_flux = [
+    {
+      desc: "Worked with the Architecture team that handled cross aspects of the project. Expansion of APIs for communicating with AAA financial entities: Coelsa, Link Network, Prisma, Banelco, others.",
+    },
+    {
+      desc: "Created functionality to retrieve and authorize using signed certificates from KeyVault, refactored code, presented an in-company .net introduction to other employees with different stacks.",
+    },
+  ];
+
+export default workExperience_flux;
