+++ home/src/components/data/workExperience_gruppoit.tsx
@@ -1,0 +1,18 @@
+const workExperience_gruppoit = [
+    { desc: "◾ Client: SmartGolf" },
+    {
+      desc: "Maintenance of web portal for Golf Community and related services. Developed new features: generation and printing process of gaming cards for the affiliated Golf clubs.",
+    },
+    { desc: " " },
+    { desc: "◾ Client: Sistema Odontológico Privado" },
+    {
+      desc: "ERP software for Odontology. Maintenance and on-site support. Added feature to display dental diagrams.",
+    },
+    { desc: " " },
+    { desc: "◾ Client: Auto Entrada" },
+    {
+      desc: "Online ticket selling portal. Generated prototypes for tickets and simple anti-forgery mechanisms (guilloches)",
+    },
+  ];
+
+export default workExperience_gruppoit;
