+++ home/src/components/data/workExperience_hka.tsx
@@ -1,0 +1,5 @@
+const workExperience_hka = [
+    { desc: "Maintained legacy Bank and Financial software." },
+  ];
+
+export default workExperience_hka;
