+++ home/src/components/data/workExperience_hp.tsx
@@ -1,0 +1,29 @@
+const workExperience_hp = [
+    { desc: "◾ Universal Music Group", bold: true },
+    {
+      desc: "Administration of legal rights of exploitation of artists and their creative work.",
+    },
+    {
+      desc: "Traveled to Los Angeles, California (USA) to be part of the maintenance from UMG.",
+    },
+    {
+      desc: "Interviewed Project Managers and programmers about how the systems worked, documenting it for later and preparing to show our understanding to the client.",
+    },
+    { desc: "Tech: .NET C#. ASP.NET. MVC. ETL processes, SQL Server, Oracle." },
+    { desc: " " },
+    { desc: "◾ Nextel" },
+    { desc: "Rebranding of 6 applications." },
+    {
+      desc: "Used J2EE and Oracle Service Bus to create, migrate and orchestrate web services.",
+    },
+    { desc: "Tech: C#, HTML, CSS, Oracle Service Bus, Oracle, and J2EE." },
+    { desc: " " },
+    { desc: "◾ American Airlines" },
+    { desc: "Maintained the web services of American Airlines." },
+    {
+      desc: "Monitored the status of the services and developed utilities and internal tools.",
+    },
+    { desc: "Tech: J2EE, Web Services, DevOps." },
+  ];
+
+export default workExperience_hp;
