+++ home/src/components/data/workExperience_registro.tsx
@@ -1,0 +1,11 @@
+const workExperience_registro = [
+    {
+      desc: "Time and attendance control application, for internal use in the General Register of the Province of Córdoba. Adapted a fingerprint reader and SDK, creating a software for time and attendance and report generation.",
+    },
+    {
+      desc: "Developed a flowchart design tool for modeling flows of documents and approvals, linked with Real State data and filling locations.",
+    },
+    { desc: "Coach junior and semi-senior developers." },
+  ];
+
+export default workExperience_registro;
