+++ home/src/components/data/workExperience_southworks.tsx
@@ -1,0 +1,12 @@
+const workExperience_southworks = [
+    {
+      desc: "Contributed to create and design a currently [NDA] undisclosable new streaming solution, for selling to Microsoft / other companies.",
+    },
+    {
+      desc: "Contributed with the integration with other APIs / live streaming platforms like YouTube, LinkedIn.",
+    },
+    { desc: "UI/UX design of the majority of the MS Teams extension." },
+    { desc: "Created functional prototype (React) for UI/UX demo." },
+  ];
+
+export default workExperience_southworks;
