+++ home/src/components/data/workExperience_tg.tsx
@@ -1,0 +1,7 @@
+const workExperience_tg = [
+    { desc: "◾ Client: Insala" },
+    { desc: "Software maintenance of legacy live project." },
+    { desc: "Rewrite and upgrade technology from scratch." },
+  ];
+
+export default workExperience_tg;
