+++ home/src/components/data/workExperience_utn.tsx
@@ -1,0 +1,16 @@
+export const pills_utn2 = [
+  { name: "J2EE" },
+  { name: "IBM DB/2" },
+  { name: "WebSphere" },
+  { name: "XML" },
+];
+
+export const workExperience_utn = [
+  { desc: "◾ Clinical Engineering Group" },
+  { desc: "Received a modest scholarship from the University." },
+  {
+    desc: "Developed a J2EE web-based system to measure technical indicators and collaborate in diagnostics of patients with potential neuropsychological problems.",
+  },
+];
+
+export default workExperience_utn;
