+++ home/src/index.css
@@ -63,6 +63,37 @@ code {
     &:focus { background: red; }
     & > span { display: inline-block; transform: skew(21deg); } /* Un-skew the text */
   }
+
+  .cutBtn {
+    clip-path: polygon(10% 0%, 100% 0, 100% 100%, 0% 100%);
+    outline: none;
+  }
+
+  .cutBtn2 {
+    clip-path: polygon(0 0, 90% 0, 100% 10%, 100% 100%, 10% 100%, 0 90%);
+    outline: none;
+  }
+
+  .parallelogram {
+    clip-path: polygon(25% 0%, 100% 0%, 75% 100%, 0% 100%);
+    outline: none;
+  }
+
+  .rightChevron {
+    clip-path: polygon(75% 0%, 100% 50%, 75% 100%, 0% 100%, 25% 50%, 0% 0%);
+    outline: none;
+  }
+  
+  .fancyDiv {
+    transition: 0.4s cubic-bezier(1, -1, 0, 2);
+    clip-path: polygon(50% 5%, 0% 100%, 100% 100%);
+    outline: none;
+  }
+  
+  .fancyDiv:hover {
+    clip-path: polygon(50% 19%, 0 76%, 100% 76%);
+    outline: none;
+  }
 }
 
 hexagon {
